assembler: rename modules after core functions
