+++ assembler/src/analyze.rs
@@ -7,10 +7,10 @@ use std::string::String;
 use itertools::{concat, Itertools, zip};
 use ariadne::{Label, Report, ReportBuilder, ReportKind};
 use lc3_isa::{Addr, SignedWord, Word};
-use crate::lexer::{LexData, LiteralValue, Opcode};
-use crate::parser::{File, get, get_result, Instruction, Operand, Region, result, WithErrData};
+use crate::lex::{LexData, LiteralValue, Opcode};
+use crate::parse::{File, get, get_result, Instruction, Operand, Region, result, WithErrData};
 use crate::{Span, Spanned, util};
-use crate::assembler::{calculate_offset, get_orig};
+use crate::assemble::{calculate_offset, get_orig};
 use crate::error::{Error, InvalidReferenceReason, OperandType, RegionPlacement, RoughAddr, SingleError};
 use crate::error::OperandType::*;
 use crate::error::Error::*;
